android: move redirect
