+++ android/App/src/main/kotlin/com/darakeon/dfm/accounts/AccountLine.kt
@@ -3,10 +3,10 @@ package com.darakeon.dfm.accounts
 import android.content.Context
 import android.util.AttributeSet
 import android.widget.LinearLayout
-import com.darakeon.dfm.lib.api.entities.accounts.Account
-import com.darakeon.dfm.extensions.redirect
-import com.darakeon.dfm.lib.extensions.setValueColored
 import com.darakeon.dfm.extract.ExtractActivity
+import com.darakeon.dfm.lib.api.entities.accounts.Account
+import com.darakeon.dfm.lib.extensions.redirect
+import com.darakeon.dfm.lib.extensions.setValueColored
 import kotlinx.android.synthetic.main.accounts_line.view.name
 import kotlinx.android.synthetic.main.accounts_line.view.value
