+++ android/App/src/main/kotlin/com/darakeon/dfm/summary/MonthLine.kt
@@ -3,9 +3,9 @@ package com.darakeon.dfm.summary
 import android.content.Context
 import android.util.AttributeSet
 import android.widget.LinearLayout
-import com.darakeon.dfm.extensions.redirect
-import com.darakeon.dfm.lib.extensions.setValueColored
 import com.darakeon.dfm.extract.ExtractActivity
+import com.darakeon.dfm.lib.extensions.redirect
+import com.darakeon.dfm.lib.extensions.setValueColored
 import kotlinx.android.synthetic.main.month_line.view.name
 import kotlinx.android.synthetic.main.month_line.view.value
